Problem - 1: Electricity Bill Calculation
Problem-01:
Design the following user interface to calculate and print bills for the city power company. The rates
depending on whether the use  is residential, commercial or industrial. The rates are computed as
follows:
Residential: 10.00 Tk. plus 0.053 Tk. per KWH used.
Commercial: 50.00 Tk. for the first 800 KWH and 0.042 Tk. for each additional KWH
Industrial: Rate varies depending on the time of usage:
Peak Hours: 70.00 Tk. for first 800 KWH and 0.062 TK for each additional KWH
Off-Peak Hour: 35.00 Tk. for first 800 KWH and 0.025 Tk. for each additional KWH